835 - Refactor and improve test utilities (#836)
* Fix errors in `TapestryApp` preventing testing of this component
* Update the test utilities module to match the @testing-library/vue API so we can hook in Vue Router
* Refactor the dataset parsing to be self contained in the dataset module (instead of being spread out in mutations and dataset) which allows progress to be initialized through Jest tests
