+++ templates/vue/src/components/settings-modal/__tests__/DuplicateTapestryButton.spec.js
@@ -19,7 +19,7 @@ describe("duplicate tapestry button", () => {
       Promise.resolve({ settings: { permalink: slug } })
     )
 
-    const screen = render(DuplicateTapestryButton, oneNodeTapestry)
+    const screen = render(DuplicateTapestryButton, { fixture: oneNodeTapestry })
     fireEvent.click(screen.getByText(/duplicate tapestry/i))
 
     expect(client.addTapestry).toHaveBeenCalledTimes(1)
